Refactor Map & Member & ShowAllAnimals pages
